+++ CHANGELOG.md
@@ -6,6 +6,11 @@ For examples and guidelines see [https://keepachangelog.com/](https://keepachang
 
 # [Unreleased]
 
+### Added
+
+- Command ´fgt hamaster´ now supports Jinja2 templating
+
+
 
 # [Released]